Moved props from Component to IOObject so other types of object can have properties too (i.e. Wires)
